Change pvp-bounds default, automate resolver selection
As described in the `README`, when using `--pvp-bounds=lower` you want
to use the lowest bounds you test with. Concretely, this often means the
`stack-lts-*.yaml` that is the earliest resolver. This can be worked out
trivially, so let's do that and stop our end-users (also us) from having
to.

It also seems reasonable for this Action to use our own preferred
defaults (when they're not super objectionable, as is this case); so
let's do that as well.