Fix broken pipe
On GitHub, if the read pulls a value from the find output, the echo
fails with a broken pipe. We can avoid this by only reading from find
then defaulting afterwards if appropriate. This has a nice benefit of
being able to hide logging if we're not actually doing anything.