Fix bug for projects with no stack-lts* files
The `read -r` expression doesn't return empty in this case, it exits
non-zero failing the step. Wrapping it in an `if` is cleaner and fixes
this bug, as confirmed in the new test case.

As a drive-by, pass inputs through `env`, a minor security best practice
to avoid interpolation of untrusted input